Improve `SearchApplication` definition (and related)
