+++ specification/search_application/get/SearchApplicationsGetResponse.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { SearchApplication } from '../_types/SearchApplication'
+import { SearchApplication } from '@search_application/_types/SearchApplication'
 
 export class Response {
   body: SearchApplication
